Implemented Feedback(Description)
First of all, Image 1, The @version tag is only allowed/used on Classes in JavaDocs
Second of All, Ive made a quick push adding, @SuppressWarnings("unused"), just removes the annoying warnings of unnused method, it doesnt add anything else
I changed canFly to True, since everyone who implements this Interface in an entity, wants their entity to fly xD
Removed setAltitude and used Entity Block Position for getAltitude
Entities have Attributes, one of which is called FLYING_SPEED, so i modified the FlyingSpeed methods to utilize that system to be compatible with other mods.
Updated the Comment regarding the FlyCooldown with
I setup a EntityStateManager which saves all the states of the entities, which the Interfaces implements, sets and returns
This allows users an already ready system to use when they want to implement these methods!